+++ packages/guideline-blocks-settings/src/components/RichTextEditor/plugins/LinkPlugin/FloatingLink/EditLinkModal/EditModal.tsx
@@ -1,43 +1,47 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-import { IconPen16, IconTrashBin16 } from '@frontify/fondue';
-import { useFloatingLinkUrlInput } from '@udecode/plate';
-import { FloatingLink } from '../FloatingLink';
+import { MouseEvent } from 'react';
+import { FloatingModalWrapper, IconPen16, IconTrashBin16, useLinkOpenButtonState } from '@frontify/fondue';
+import { getUrlFromLinkOrLegacyLink } from '../../../../../Link';
 
-export const EditModal = () => {
-    const urlHtmlProps = useFloatingLinkUrlInput({});
+type EditModalProps = {
+    editButtonProps: {
+        onClick: () => void;
+    };
+    unlinkButtonProps: {
+        onMouseDown: (e: MouseEvent<HTMLButtonElement>) => void;
+        onClick: () => void;
+    };
+};
+
+export const EditModal = ({ editButtonProps, unlinkButtonProps }: EditModalProps) => {
+    const { element } = useLinkOpenButtonState();
+    const url = element ? getUrlFromLinkOrLegacyLink(element) : '';
 
     return (
-        <div
-            data-test-id="floating-link-edit"
-            className="tw-bg-white tw-text-text tw-rounded tw-shadow tw-p-4 tw-min-w-[400px]"
-        >
-            <span data-test-id={'preview-link-flyout'} className="tw-flex tw-justify-between tw-items-center">
-                <span className="tw-pointer-events-none">{urlHtmlProps.defaultValue}</span>
+        <FloatingModalWrapper data-test-id="floating-link-edit" padding="16px" minWidth="400px">
+            <span data-test-id="preview-link-flyout" className="tw-flex tw-justify-between tw-items-center">
+                <span className="tw-pointer-events-none">{url}</span>
                 <span className="tw-flex tw-gap-2">
-                    <span
-                        role="button"
+                    <button
                         tabIndex={0}
-                        data-test-id={'edit-link-button'}
+                        data-test-id="edit-link-button"
                         className="tw-transition tw-cursor-pointer tw-rounded hover:tw-bg-black-10 tw-p-1"
+                        {...editButtonProps}
                     >
-                        <FloatingLink.EditButton>
-                            <IconPen16 />
-                        </FloatingLink.EditButton>
-                    </span>
+                        <IconPen16 />
+                    </button>
 
-                    <span
-                        role="button"
+                    <button
                         tabIndex={0}
-                        data-test-id={'remove-link-button'}
+                        data-test-id="remove-link-button"
                         className="tw-transition tw-cursor-pointer tw-rounded hover:tw-bg-black-10 tw-p-1"
+                        {...unlinkButtonProps}
                     >
-                        <FloatingLink.UnlinkButton>
-                            <IconTrashBin16 />
-                        </FloatingLink.UnlinkButton>
-                    </span>
+                        <IconTrashBin16 />
+                    </button>
                 </span>
             </span>
-        </div>
+        </FloatingModalWrapper>
     );
 };