+++ packages/guideline-blocks-settings/src/components/RichTextEditor/plugins/LinkPlugin/types.ts
@@ -1,6 +1,6 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-import { TLinkElement as TPlateLinkElement } from '@udecode/plate';
+import { TLinkElement as TPlateLinkElement } from '@frontify/fondue';
 
 export type TLinkElement = TPlateLinkElement & {
     chosenLink?: {